+++ packages/mastodon/lib/src/api/accounts_api.dart
@@ -60,7 +60,8 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
-          {'type': 'oauth2', 'name': 'OAuth2ClientCredentials'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -593,6 +594,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -677,6 +679,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -755,6 +758,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -835,6 +839,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -928,6 +933,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1039,6 +1045,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1209,6 +1216,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1289,6 +1297,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1368,6 +1377,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1468,6 +1478,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1551,6 +1562,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1637,6 +1649,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1739,6 +1752,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1840,6 +1854,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1940,6 +1955,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -2023,6 +2039,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -2108,6 +2125,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -2191,6 +2209,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -2276,6 +2295,7 @@ class AccountsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -2359,6 +2379,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -2444,6 +2465,7 @@ class AccountsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,