+++ packages/mastodon/lib/src/api/apps_api.dart
@@ -136,7 +136,8 @@ class AppsApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
-          {'type': 'oauth2', 'name': 'OAuth2ClientCredentials'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },