+++ packages/mastodon/lib/src/api/domain_blocks_api.dart
@@ -49,6 +49,7 @@ class DomainBlocksApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -112,6 +113,7 @@ class DomainBlocksApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -182,6 +184,7 @@ class DomainBlocksApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,