+++ packages/mastodon/lib/src/api/filters_api.dart
@@ -58,6 +58,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -151,6 +152,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -248,6 +250,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -301,6 +304,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -356,6 +360,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -409,6 +414,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -495,6 +501,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -578,6 +585,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -663,6 +671,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -747,6 +756,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -820,6 +830,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -905,6 +916,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -989,6 +1001,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1065,6 +1078,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1151,6 +1165,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1252,6 +1267,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
@@ -1355,6 +1371,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1456,6 +1473,7 @@ class FiltersApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },
@@ -1552,6 +1570,7 @@ class FiltersApi {
       headers: <String, dynamic>{...?headers},
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
+          {'type': 'oauth2', 'name': 'OAuth2'},
           {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,