+++ packages/mastodon/lib/src/api/followed_tags_api.dart
@@ -54,6 +54,7 @@ class FollowedTagsApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },