+++ packages/mastodon/lib/src/api/oauth_api.dart
@@ -122,6 +122,7 @@ class OauthApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },