+++ packages/mastodon/lib/src/api/preferences_api.dart
@@ -43,6 +43,7 @@ class PreferencesApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },