+++ packages/mastodon/lib/src/api/search_api.dart
@@ -66,6 +66,7 @@ class SearchApi {
       extra: <String, dynamic>{
         'secure': <Map<String, String>>[
           {'type': 'oauth2', 'name': 'OAuth2'},
+          {'type': 'oauth2', 'name': 'OAuth2'},
         ],
         ...?extra,
       },